requirements.txt
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pyjson==1.3.0
PyYAML==6.0
Werkzeug==2.2.2
requests==2.28.1
